+++ rdagent/components/coder/CoSTEER/knowledge_management.py
@@ -6,7 +6,7 @@ import random
 import re
 from itertools import combinations
 from pathlib import Path
-from typing import Union
+from typing import List, Union
 
 from jinja2 import Environment, StrictUndefined
 
@@ -339,6 +339,7 @@ class CoSTEERRAGStrategyV2(RAGStrategy):
                     system_prompt=analyze_component_system_prompt,
                     user_prompt=analyze_component_user_prompt,
                     json_mode=True,
+                    json_target_type=List[int],
                 ),
             )["component_no_list"]
             return [all_component_nodes[index - 1] for index in sorted(list(set(component_no_list)))]
