+++ rdagent/scenarios/data_mining/developer/feedback.py
@@ -3,6 +3,7 @@
 
 import json
 from pathlib import Path
+from typing import Dict
 
 from jinja2 import Environment, StrictUndefined
 
@@ -59,6 +60,7 @@ class DMModelExperiment2Feedback(Experiment2Feedback):
             user_prompt=user_prompt,
             system_prompt=system_prompt,
             json_mode=True,
+            json_target_type=Dict[str, str | bool | int],
         )
 
         # Parse the JSON response to extract the feedback
