+++ rdagent/scenarios/data_science/scen/__init__.py
@@ -1,6 +1,7 @@
 import json
 import os
 from pathlib import Path
+from typing import Dict
 
 import pandas as pd
 from PIL import Image, TiffTags
@@ -258,6 +259,7 @@ class DataScienceScen(Scenario):
             user_prompt=user_prompt,
             system_prompt=sys_prompt,
             json_mode=True,
+            json_target_type=Dict[str, str | int | bool],
         )
 
         response_json_analysis = json.loads(response_analysis)
