+++ rdagent/scenarios/kaggle/developer/feedback.py
@@ -1,5 +1,6 @@
 import json
 from pathlib import Path
+from typing import Dict
 
 import pandas as pd
 from jinja2 import Environment, StrictUndefined
@@ -148,6 +149,7 @@ class KGExperiment2Feedback(Experiment2Feedback):
             user_prompt=usr_prompt,
             system_prompt=sys_prompt,
             json_mode=True,
+            json_target_type=Dict[str, str | bool | int],
         )
 
         response_json = json.loads(response)
