+++ rdagent/scenarios/kaggle/experiment/scenario.py
@@ -3,6 +3,7 @@ import json
 import pickle
 from datetime import datetime, timezone
 from pathlib import Path
+from typing import Dict
 
 import pandas as pd
 from jinja2 import Environment, StrictUndefined
@@ -93,6 +94,7 @@ class KGScenario(Scenario):
             user_prompt=user_prompt,
             system_prompt=sys_prompt,
             json_mode=True,
+            json_target_type=Dict[str, str | bool | int],
         )
 
         response_json_analysis = json.loads(response_analysis)
